BusyBox/Alpine support
The BusyBox stat and touch binaries differ from their GNU counterparts
in some ways that are relevant to rig:

The stat binary won't accept the --printf flag, but it accepts the
-c flag in the same way GNU stat does. The -c command won't accept
backslash escapes, though. Hence fall back to the pipe character as
delimiter.

The touch binary won't parse timezones and sub-second precision times.
Accomodate for that by normalizing the time to UTC and only use sub-
second precision timestamps if it's not BusyBox touch. This will loose
precision, but that's probably better than failing completely.

Make localhost connections use the SHELL instead of hardcoding bash.
Make the rigfs helper shell script POSIX compatible.

Add an OS support package for Alpine that uses apk for package
installations.

Add a minimal footloose-compatible Dockerfile for Alpine 3.18 to the
tests and execute it as part of the CI.